# docs/swagger.yaml
basePath: /
definitions:
...
  models.Deals:
    properties:
      amount:
        type: string
      created_at:
        type: string
      currency:
        type: string
      id:
        type: integer
      lead_id:
        type: integer
      status:
        type: string
    type: object
  models.Document:
    properties:
      deal_id:
        type: integer
      doc_type:
        type: string
      file_path:
        type: string
      id:
        type: integer
      signed_at:
        type: string
      status:
        type: string
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@company.com
    name: Тур Компани
    url: http://contact.company.com
  description: Aпишка для управления туристической компанией (документация Swagger).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://your-terms.com/terms/
  title: Tour Company API
  version: "1.0"
paths:
  /deals:
    post:
      consumes:
      - application/json
      description: Создает новую сделку, связанную с лидом
      parameters:
      - description: Данные сделки
        in: body
        name: deals
        required: true
        schema:
          $ref: '#/definitions/models.Deals'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Deals'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создание сделки
      tags:
      - Deals
  /deals/:
    get:
      description: Returns a list of all deals
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Deals'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all deals
      tags:
      - Deals
  /deals/{id}:
    delete:
      description: Удаляет сделку по ID
      parameters:
      - description: ID сделки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить сделку
      tags:
      - Deals
    get:
      description: Возвращает данные одной сделки
      parameters:
      - description: ID сделки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Deals'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить сделку по ID
      tags:
      - Deals
    put:
      consumes:
      - application/json
      description: Обновляет данные сделки по ее ID.
      parameters:
      - description: ID сделки
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные сделки
        in: body
        name: deal
        required: true
        schema:
          $ref: '#/definitions/models.Deals'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Deals'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновление сделки
      tags:
      - Deals
  /documents:
    post:
      consumes:
      - application/json
      description: Создает новый документ
      parameters:
      - description: Данные документа
        in: body
        name: document
        required: true
        schema:
          $ref: '#/definitions/models.Document'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создание документа
      tags:
      - Documents
  /documents/{id}:
    delete:
      description: Удаляет документ по ID
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить документ
      tags:
      - Documents
    get:
      description: Возвращает один документ по его ID
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Document'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить документ по ID
      tags:
      - Documents
  /documents/deal/{dealid}:
    get:
      description: Возвращает все документы, связанные с определенной сделкой
      parameters:
      - description: ID сделки
        in: path
        name: dealid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Document'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить документы сделки
      tags:
      - Documents
  /documents/from-lead:
    post:
      consumes:
      - application/json
      description: Создает документ на основе лида и типа документа
      parameters:
      - description: ID лида и тип документа
        in: body
        name: input
        required: true
        schema:
          properties:
            doc_type:
              type: string
            lead_id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создание документа из лида
      tags:
      - Documents
...
swagger: "2.0"